fix(telemetry): emit instance_started/snapshot_history/exited from main
These four events fired from the panel renderer's onInstanceStarted /
onComfyExited callbacks, but Desktop 2's unified window tears the dashboard
panel down around server-ready/exit, so the callbacks frequently never ran
and the events vanished from PostHog after rc.1 (instance_started,
installation_started, snapshot_history) and were nearly absent for exited.

Move emission to the main process where it fires reliably regardless of
renderer lifecycle, preserving the exact event names and property shapes so
existing dashboards/SQL keep working:

- shared.ts: extract buildInstallationDdContext (reused by the IPC handler);
  add an onInstanceStarted callback fired from _addSession with the same
  boot-time/retry signal that drove the instance-started broadcast.
- sessionStartTelemetry.ts (new): main-side emitter for instance_started,
  the deprecated installation_started shadow, and snapshot_history, with
  emit-site PII scrubbing of pip specs / node ids.
- launch.ts: emit comfy.desktop.comfyui.exited via telemetry.emit (PostHog +
  Datadog crash-rate monitor) from both exit handlers.
- rendererBootstrap.ts: drop the now-duplicate renderer emissions (keep
  boot_log, which still fires while the panel is alive).

Amp-Thread-ID: https://ampcode.com/threads/T-019ef644-0d40-7467-b80b-90f8e7b2cd3e
