+++ src/main/index.ts
@@ -51,6 +51,7 @@ import { hasActiveTemplateDownloads, getTemplateDownloadState } from './sources/
 import { isTerminal as isTemplateDownloadTerminal } from './sources/standalone/templateDownloadCore'
 import { registerAssetDownloadHandlers } from './lib/ipc/registerAssetDownloadHandlers'
 import { registerDownloadHandlers } from './lib/ipc/registerDownloadHandlers'
+import { emitInstanceStartedTelemetry } from './lib/ipc/sessionStartTelemetry'
 import {
   get as getInstallation,
   installationEvents,
@@ -2039,6 +2040,9 @@ if (app.isPackaged && !app.requestSingleInstanceLock()) {
       onLaunch,
       onStop,
       onComfyExited,
+      onInstanceStarted: (info) => {
+        void emitInstanceStartedTelemetry(info)
+      },
       onComfyRestarted,
       onModelFolderRelaunch,
       onLocaleChanged: updateTrayMenu,
